+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MinHealthPercent` to spell/skill metadata
 - `MaxHealthPercent` to spell/skill metadata (Crasher/Animal Feast/Execute)
 - `OpensDialog` to spell metadata (was only skills before)
+- More tooltips in metadata editors
 
 ### Changed
 
 - Use new metadata properties for HP based conditions
+- Redesign some metadata editors for skill/spells
 - Increase dialog size for skill/spell metadata editors
 
 ### Removed